+++ lib/authentication.dart
@@ -1,4 +1,4 @@
-library authentication;
+library;
 
 export 'src/authentication/abstract_authentication.dart';
 export 'src/authentication/cra_authentication.dart';