+++ lib/cbor.dart
@@ -1,3 +1,3 @@
-library cbor;
+library;
 
 export 'src/serializer/cbor/serializer.dart';