+++ lib/json.dart
@@ -1,3 +1,3 @@
-library json;
+library;
 
 export 'src/serializer/json/serializer.dart';