+++ lib/socket.dart
@@ -1,4 +1,4 @@
-library socket;
+library;
 
 export 'src/transport/socket/socket_helper.dart';
 export 'src/transport/socket/socket_transport.dart';